refactor(hooks): move generate_team_name + session-id validity predicate to shared.pact_context (#943)
